Convert to a zoned-time from a naive-time

Converting to a zoned-time from a naive-time retains the printed time, but changes the underlying duration, depending on the zone that you choose.

Naive-times are time points with a yet-to-be-determined time zone. By converting them to a zoned-time, all you are doing is specifying that time zone while attempting to keep all other printed information the same (if possible).

If you want to retain the underlying duration, try converting to a zoned-time from a sys-time, which is a time point interpreted as having a UTC time zone.

# S3 method for clock_naive_time
as_zoned_time(x, zone, ..., nonexistent = "error", ambiguous = "error")

Arguments

x
[clock_naive_time]

A naive-time to convert to a zoned-time.

zone
[character(1)]

The zone to convert to.

...
These dots are for future extensions and must be empty.

nonexistent
[character]

One of the following nonexistent time resolution strategies:

"roll-forward": The next valid instant in time.

"roll-backward": The previous valid instant in time.

"shift-forward": Shift the nonexistent time forward by the size of the daylight saving time gap.

"shift-backward: Shift the nonexistent time backward by the size of the daylight saving time gap.

"NA": Replace nonexistent times with NA.

"error": Error on nonexistent times.

Allowed to be either length 1, or the same length as the input.

Using either "roll-forward" or "roll-backward" is generally recommended over shifting, as these two strategies maintain the relative ordering between elements of the input.

ambiguous
[character]

One of the following ambiguous time resolution strategies:

"earliest": Of the two possible times, choose the earliest one.

"latest": Of the two possible times, choose the latest one.

"NA": Replace ambiguous times with NA.

"error": Error on ambiguous times.

Allowed to be either length 1, or the same length as the input.

Value

A zoned-time vector.

Daylight Saving Time

Converting from a naive-time to a zoned-time is not always possible due to daylight saving time issues. There are two types of these issues:

Nonexistent times are the result of daylight saving time "gaps". For example, in the America/New_York time zone, there was a daylight saving time gap 1 second after "2020-03-08 01:59:59", where the clocks changed from 01:59:59 -> 03:00:00, completely skipping the 2 o'clock hour. This means that if you had a naive time of "2020-03-08 02:30:00", you couldn't convert that straight into a zoned-time with this time zone. To resolve these issues, the nonexistent argument can be used to specify one of many nonexistent time resolution strategies.

Ambiguous times are the result of daylight saving time "fallbacks". For example, in the America/New_York time zone, there was a daylight saving time fallback 1 second after "2020-11-01 01:59:59 EDT", at which point the clocks "fell backwards" by 1 hour, resulting in a printed time of "2020-11-01 01:00:00 EST" (note the EDT->EST shift). This resulted in two 1 o'clock hours for this day, so if you had a naive time of "2020-11-01 01:30:00", you wouldn't be able to convert that directly into a zoned-time with this time zone, as there is no way for clock to know which of the two ambiguous times you wanted. To resolve these issues, the ambiguous argument can be used to specify one of many ambiguous time resolution strategies.

Examples

x <- as_naive_time(year_month_day(2019, 1, 1))

# Converting a naive-time to a zoned-time generally retains the
# printed time, while changing the underlying duration.
as_zoned_time(x, "America/New_York")

#> <zoned_time<second><America/New_York>[1]>
#> [1] "2019-01-01T00:00:00-05:00"

as_zoned_time(x, "America/Los_Angeles")

#> <zoned_time<second><America/Los_Angeles>[1]>
#> [1] "2019-01-01T00:00:00-08:00"

# ---------------------------------------------------------------------------
# Nonexistent time:

new_york <- "America/New_York"

# There was a daylight saving gap in the America/New_York time zone on
# 2020-03-08 01:59:59 -> 03:00:00, which means that one of these
# naive-times don't exist in that time zone. By default, attempting to
# convert it to a zoned time will result in an error.
nonexistent_time <- year_month_day(2020, 03, 08, c(02, 03), c(45, 30), 00)
nonexistent_time <- as_naive_time(nonexistent_time)
try(as_zoned_time(nonexistent_time, new_york))

#> Error : Nonexistent time due to daylight savings at location 1. Resolve nonexistent time issues by specifying the `nonexistent` argument.

# Resolve this by specifying a nonexistent time resolution strategy
as_zoned_time(nonexistent_time, new_york, nonexistent = "roll-forward")

#> <zoned_time<second><America/New_York>[2]>
#> [1] "2020-03-08T03:00:00-04:00" "2020-03-08T03:30:00-04:00"

as_zoned_time(nonexistent_time, new_york, nonexistent = "roll-backward")

#> <zoned_time<second><America/New_York>[2]>
#> [1] "2020-03-08T01:59:59-05:00" "2020-03-08T03:30:00-04:00"

# Note that rolling backwards will choose the last possible moment in
# time at the current precision of the input
nonexistent_nanotime <- time_point_cast(nonexistent_time, "nanosecond")
nonexistent_nanotime

#> <time_point<naive><nanosecond>[2]>
#> [1] "2020-03-08 02:45:00.000000000" "2020-03-08 03:30:00.000000000"

as_zoned_time(nonexistent_nanotime, new_york, nonexistent = "roll-backward")

#> <zoned_time<nanosecond><America/New_York>[2]>
#> [1] "2020-03-08T01:59:59.999999999-05:00" "2020-03-08T03:30:00.000000000-04:00"

# A word of caution - Shifting does not guarantee that the relative ordering
# of the input is maintained
shifted <- as_zoned_time(nonexistent_time, new_york, nonexistent = "shift-forward")
shifted

#> <zoned_time<second><America/New_York>[2]>
#> [1] "2020-03-08T03:45:00-04:00" "2020-03-08T03:30:00-04:00"

# 02:45:00 < 03:30:00
nonexistent_time[1] < nonexistent_time[2]

#> [1] TRUE

# 03:45:00 > 03:30:00 (relative ordering is lost)
shifted[1] < shifted[2]

#> [1] FALSE

# ---------------------------------------------------------------------------
# Ambiguous time:

new_york <- "America/New_York"

# There was a daylight saving time fallback in the America/New_York time
# zone on 2020-11-01 01:59:59 EDT -> 2020-11-01 01:00:00 EST, resulting
# in two 1 o'clock hours. This means that the following naive time is
# ambiguous since we don't know which of the two 1 o'clocks it belongs to.
# By default, attempting to convert it to a zoned time will result in an
# error.
ambiguous_time <- year_month_day(2020, 11, 01, 01, 30, 00)
ambiguous_time <- as_naive_time(ambiguous_time)
try(as_zoned_time(ambiguous_time, new_york))

#> Error : Ambiguous time due to daylight savings at location 1. Resolve ambiguous time issues by specifying the `ambiguous` argument.

# Resolve this by specifying an ambiguous time resolution strategy
as_zoned_time(ambiguous_time, new_york, ambiguous = "earliest")

#> <zoned_time<second><America/New_York>[1]>
#> [1] "2020-11-01T01:30:00-04:00"

as_zoned_time(ambiguous_time, new_york, ambiguous = "latest")

#> <zoned_time<second><America/New_York>[1]>
#> [1] "2020-11-01T01:30:00-05:00"

as_zoned_time(ambiguous_time, new_york, ambiguous = "NA")

#> <zoned_time<second><America/New_York>[1]>
#> [1] NA