as_weekday {clock}	R Documentation

Convert to a weekday

Description

as_weekday() converts to a weekday type. This is normally useful for converting to a weekday from a sys-time or naive-time. You can use this function along with the circular arithmetic that weekday implements to easily get to the "next Monday" or "previous Sunday".

Usage

as_weekday(x)

Arguments

x
[object]

An object to convert to a weekday. Usually a sys-time or naive-time.

Value

A weekday.

Examples

x <- as_naive_time(year_month_day(2019, 01, 05))

# This is a Saturday!
as_weekday(x)

# See the examples in `?weekday` for more usage.

[Package clock version 0.3.0 Index]