as_zoned_time {clock}	R Documentation

Convert to a zoned-time

Description

as_zoned_time() converts x to a zoned-time. You generally convert to a zoned time from either a sys-time or a naive time. Each are documented on their own page:

sys-time

naive-time

There are also convenience methods for converting to a zoned time from native R date and date-time types:

dates (Date)

date-times (POSIXct / POSIXlt)

Usage

as_zoned_time(x, ...)

Arguments

x
[object]

An object to convert to a zoned-time.

...
These dots are for future extensions and must be empty.

Value

A zoned-time vector.

Examples

x <- as.Date("2019-01-01")
as_zoned_time(x, "Europe/London")

y <- as_naive_time(year_month_day(2019, 2, 1))
as_zoned_time(y, zone = "America/New_York")

[Package clock version 0.3.0 Index]