calendar_month_factor {clock}	R Documentation

Convert a calendar to an ordered factor of month names

Description

calendar_month_factor() extracts the month values from a calendar and converts them to an ordered factor of month names. This can be useful in combination with ggplot2, or for modeling.

This function is only relevant for calendar types that use a month field, i.e. year_month_day() and year_month_weekday(). The calendar type must have at least month precision.

Usage

calendar_month_factor(x, ..., labels = "en", abbreviate = FALSE)

Arguments

x
[calendar]

A calendar vector.

...
These dots are for future extensions and must be empty.

labels
[clock_labels / character(1)]

Character representations of localized weekday names, month names, and AM/PM names. Either the language code as string (passed on to clock_labels_lookup()), or an object created by clock_labels().

abbreviate
[logical(1)]

If TRUE, the abbreviated month names from labels will be used.

If FALSE, the full month names from labels will be used.

Value

An ordered factor representing the months.

Examples

x <- year_month_day(2019, 1:12)

calendar_month_factor(x)
calendar_month_factor(x, abbreviate = TRUE)
calendar_month_factor(x, labels = "fr")

[Package clock version 0.3.0 Index]