Date-time arithmetic

Date-time arithmetic in clock is broken down into two types: zoned and naive. Each type is documented on its own help page.

Zoned dates have a time zone attached. The two base R classes, POSIXct and Date, implement zoned dates (Date is assumed to implicitly have a time zone of UTC).

Naive dates are independent of a time zone.

add_years(x, n, ...)

subtract_years(x, n, ...)

add_quarters(x, n, ...)

subtract_quarters(x, n, ...)

add_months(x, n, ...)

subtract_months(x, n, ...)

add_weeks(x, n, ...)

subtract_weeks(x, n, ...)

add_days(x, n, ...)

subtract_days(x, n, ...)

add_hours(x, n, ...)

subtract_hours(x, n, ...)

add_minutes(x, n, ...)

subtract_minutes(x, n, ...)

add_seconds(x, n, ...)

subtract_seconds(x, n, ...)

add_milliseconds(x, n, ...)

subtract_milliseconds(x, n, ...)

add_microseconds(x, n, ...)

subtract_microseconds(x, n, ...)

add_nanoseconds(x, n, ...)

subtract_nanoseconds(x, n, ...)

Arguments

x
[Date / POSIXct / POSIXlt]

A zoned or naive date-time vector.

n
[integer]

An integer vector representing the number of units to add to or subtract from x.

...
These dots are for future extensions and must be empty.