Naive date-time arithmetic

Naive arithmetic involves adding or subtracting units of time from a datetime that is independent of any time zone. This means that daylight savings time is never an issue while working with a naive datetime. Usually, you will convert to a naive datetime with as_naive_time_point(), perform multiple arithmetic operations or adjustments with it, then convert it back to a zoned datetime with as_zoned_time_point().

Naive arithmetic is usually appropriate when performing multiple arithmetic operations in a row (like adding a set of years, months, and days). The alternative is zoned arithmetic, which is simpler and more straightforward when you just need to perform a single operation (like just adding 3 months).

# S3 method for clock_naive_time_point
subtract_years(x, n, ..., day_nonexistent = "last-time")

# S3 method for clock_calendar
subtract_years(x, n, ..., day_nonexistent = "last-time")

# S3 method for clock_naive_time_point
subtract_quarters(x, n, ..., day_nonexistent = "last-time")

# S3 method for clock_calendar
subtract_quarters(x, n, ..., day_nonexistent = "last-time")

# S3 method for clock_naive_time_point
subtract_months(x, n, ..., day_nonexistent = "last-time")

# S3 method for clock_calendar
subtract_months(x, n, ..., day_nonexistent = "last-time")

# S3 method for clock_naive_time_point
subtract_weeks(x, n, ...)

# S3 method for clock_calendar
subtract_weeks(x, n, ...)

# S3 method for clock_naive_time_point
subtract_days(x, n, ...)

# S3 method for clock_calendar
subtract_days(x, n, ...)

# S3 method for clock_naive_time_point
add_hours(x, n, ...)

# S3 method for clock_calendar
add_hours(x, n, ...)

# S3 method for clock_naive_time_point
subtract_hours(x, n, ...)

# S3 method for clock_calendar
subtract_hours(x, n, ...)

# S3 method for clock_naive_time_point
add_minutes(x, n, ...)

# S3 method for clock_calendar
add_minutes(x, n, ...)

# S3 method for clock_naive_time_point
subtract_minutes(x, n, ...)

# S3 method for clock_calendar
subtract_minutes(x, n, ...)

# S3 method for clock_naive_time_point
add_seconds(x, n, ...)

# S3 method for clock_calendar
add_seconds(x, n, ...)

# S3 method for clock_naive_time_point
subtract_seconds(x, n, ...)

# S3 method for clock_calendar
subtract_seconds(x, n, ...)

# S3 method for clock_naive_time_point
add_milliseconds(x, n, ...)

# S3 method for clock_calendar
add_milliseconds(x, n, ...)

# S3 method for clock_naive_time_point
subtract_milliseconds(x, n, ...)

# S3 method for clock_calendar
subtract_milliseconds(x, n, ...)

# S3 method for clock_naive_time_point
add_microseconds(x, n, ...)

# S3 method for clock_calendar
add_microseconds(x, n, ...)

# S3 method for clock_naive_time_point
subtract_microseconds(x, n, ...)

# S3 method for clock_calendar
subtract_microseconds(x, n, ...)

# S3 method for clock_naive_time_point
add_nanoseconds(x, n, ...)

# S3 method for clock_calendar
add_nanoseconds(x, n, ...)

# S3 method for clock_naive_time_point
subtract_nanoseconds(x, n, ...)

# S3 method for clock_calendar
subtract_nanoseconds(x, n, ...)

Arguments

x
[clock_naive]

A naive date-time vector.

n
[integer]

An integer vector representing the number of units to add to or subtract from x.

...
These dots are for future extensions and must be empty.

day_nonexistent
[character(1)]

Control the behavior when a nonexistent day is generated.

"last-time": Adjust to the last possible time of the current month.

"first-time": Adjust to the first possible time of the following month.

"last-day": Adjust to the last day of the current month. For date-times, the sub-daily components are kept.

"first-day": Adjust to the first day of the following month. For date-times, the sub-daily components are kept.

"NA": Replace the nonexistent date with NA.

"error": Error on nonexistent dates.

Warning: When used in arithmetic with date-times, "last-day" and "first-day" do not guarantee that the relative ordering of x is maintained.

Nonexistent Days

Naive datetimes are unique because they are allowed to land on nonexistent days. These might occur from adding 1 month to "1971-01-29", which theoretically lands on "1971-02-29", a nonexistent day. With zoned arithmetic, you are immediately forced to make a decision on how to handle this nonexistent day. With naive arithmetic, that decision is delayed, allowing you precise control over how to handle this case. As an example, you might choose to add 1 year to this nonexistent date, resulting in "1972-02-29", which does exist due to it being a leap year. Or you might convert it back to a zoned date with as_zoned_time_point(), which forces you to deal with the nonexistent date using the day_nonexistent argument. The default would choose the last real day in February of that year, resulting in "1971-02-28".

There are operations that force you to resolve nonexistent days. Using add_days(), add_hours(), add_minutes(), or add_seconds() will force you to resolve the nonexistent day by using the day_nonexistent argument. After the nonexistent day has been resolved, the unit of time is added.