clock-package {clock}	R Documentation

clock: Date-Time Types and Tools

Description

Provides a comprehensive library for date-time manipulations using a new family of orthogonal date-time classes (durations, time points, zoned-times, and calendars) that partition responsibilities so that the complexities of time zones are only considered when they are really needed. Capabilities include: date-time parsing, formatting, arithmetic, extraction and updating of components, and rounding.

Author(s)

Maintainer: Davis Vaughan davis@rstudio.com

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://clock.r-lib.org

https://github.com/r-lib/clock

Report bugs at https://github.com/r-lib/clock/issues

[Package clock version 0.3.0 Index]