date-today {clock}	R Documentation

Current date and date-time

Description

date_today() returns the current date in the specified zone as a Date.

date_now() returns the current date-time in the specified zone as a POSIXct.

Usage

date_today(zone)

date_now(zone)

Arguments

zone
[character(1)]

A time zone to get the current time for.

Details

clock assumes that Date is a naive type, like naive-time. This means that date_today() first looks up the current date-time in the specified zone, then converts that to a Date, retaining the printed time while dropping any information about that time zone.

Value

date_today() a single Date.

date_now() a single POSIXct.

Examples

# Current date in the local time zone
date_today("")

# Current date in a specified time zone
date_today("Europe/London")

# Current date-time in that same time zone
date_now("Europe/London")

[Package clock version 0.3.0 Index]