Parse into a year-month-day calendar

parse_year_month_day() parses a character vector into a year-month-day calendar.

parse_year_month_day(x, ..., precision = "day")

Arguments

x
[character]

A character vector to parse.

...
These dots are for future extensions and must be empty.

precision
[character(1)]

The precision of the resulting year-month-day.

Value

A year-month-day calendar vector. If a parsing fails, NA is returned.

Details

There is no format argument, as this function can only parse the result of format()-ting a year-month-day. If you need the flexibility of a format argument, try parse_naive_time() or parse_sys_time().

Examples

x <- "2019-01-01"

# Default parses at day precision
parse_year_month_day(x)

#> <year_month_day<day>[invalid=0][1]>
#> [1] "2019-01-01"

# Can parse at less precise precisions too
parse_year_month_day(x, precision = "month")

#> <year_month_day<month>[invalid=0][1]>
#> [1] "2019-01"

parse_year_month_day(x, precision = "year")

#> <year_month_day<year>[invalid=0][1]>
#> [1] "2019"

# Even invalid dates can be round-tripped through format<->parse calls
invalid <- year_month_day(2019, 2, 30)
parse_year_month_day(format(invalid))

#> <year_month_day<day>[invalid=1][1]>
#> [1] "2019-02-30"

# Can parse with time of day
x <- year_month_day(
  2019, 01, 30, 02, 30, 00, 5000,
  subsecond_precision = "nanosecond"
)

parse_year_month_day(format(x), precision = "nanosecond")

#> <year_month_day<nanosecond>[invalid=0][1]>
#> [1] "2019-01-30 02:30:00.000005000"