posixt-getters {clock}	R Documentation

Getters: date-time

Description

These are POSIXct/POSIXlt methods for the getter generics.

get_year() returns the Gregorian year.

get_month() returns the month of the year.

get_day() returns the day of the month.

There are sub-daily getters for extracting more precise components, up to a precision of seconds.

For more advanced component extraction, convert to the calendar type that you are interested in.

Usage

## S3 method for class 'POSIXt'
get_year(x)

## S3 method for class 'POSIXt'
get_month(x)

## S3 method for class 'POSIXt'
get_day(x)

## S3 method for class 'POSIXt'
get_hour(x)

## S3 method for class 'POSIXt'
get_minute(x)

## S3 method for class 'POSIXt'
get_second(x)

Arguments

x
[POSIXct / POSIXlt]

A date-time type to get the component from.

Value

The component.

Examples

x <- as.POSIXct("2019-01-01", tz = "America/New_York")

x <- add_days(x, 0:5)
x <- set_second(x, 10:15)

get_day(x)
get_second(x)

[Package clock version 0.3.0 Index]