sys_time_now {clock}	R Documentation

What is the current sys-time?

Description

sys_time_now() returns the current time in UTC.

Usage

sys_time_now()

Details

The time is returned with a nanosecond precision, but the actual amount of data returned is OS dependent. Usually, information at at least the microsecond level is returned, with some platforms returning nanosecond information.

Value

A sys-time of the current time in UTC.

Examples

x <- sys_time_now()

[Package clock version 0.3.0 Index]