Extract weekday encodings

Given a weekday object, x:

weekday_c_encoding() returns the underlying C encoding, with a range of [0, 6], where 0 = Sunday, and 6 = Saturday.

weekday_iso_encoding() returns the underlying ISO encoding, with a range of [1, 7], where 1 = Monday, and 7 = Sunday.

weekday_c_encoding(x)

weekday_iso_encoding(x)

Arguments

x
[weekday]

A weekday vector.

Value

An integer vector of encodings.

Examples

# Here we supply a C encoding to `weekday()`
x <- weekday(0:6)
x

#> <weekday[7]>
#> [1] Sun Mon Tue Wed Thu Fri Sat

# Notice that Sunday is the only day that is different
weekday_c_encoding(x)

#> [1] 0 1 2 3 4 5 6

weekday_iso_encoding(x)

#> [1] 7 1 2 3 4 5 6