weekday_code {clock}	R Documentation

Extract underlying weekday codes

Description

weekday_code() extracts out the integer code for the weekday.

Usage

weekday_code(x, ..., encoding = "western")

Arguments

x
[weekday]

A weekday vector.

...
These dots are for future extensions and must be empty.

encoding
[character(1)]

One of:

"western": Encode weekdays assuming 1 == Sunday and 7 == Saturday.

"iso": Encode weekdays assuming 1 == Monday and 7 == Sunday. This is in line with the ISO standard.

Value

An integer vector of codes.

Examples

# Here we supply a western encoding to `weekday()`
x <- weekday(1:7)
x

# We can extract out the codes using different encodings
weekday_code(x, encoding = "western")
weekday_code(x, encoding = "iso")

[Package clock version 0.3.0 Index]