year-month-day-setters {clock}	R Documentation

Setters: year-month-day

Description

These are year-month-day methods for the setter generics.

set_year() sets the Gregorian year.

set_month() sets the month of the year. Valid values are in the range of [1, 12].

set_day() sets the day of the month. Valid values are in the range of [1, 31].

There are sub-daily setters for setting more precise components.

Usage

## S3 method for class 'clock_year_month_day'
set_year(x, value, ...)

## S3 method for class 'clock_year_month_day'
set_month(x, value, ...)

## S3 method for class 'clock_year_month_day'
set_day(x, value, ...)

## S3 method for class 'clock_year_month_day'
set_hour(x, value, ...)

## S3 method for class 'clock_year_month_day'
set_minute(x, value, ...)

## S3 method for class 'clock_year_month_day'
set_second(x, value, ...)

## S3 method for class 'clock_year_month_day'
set_millisecond(x, value, ...)

## S3 method for class 'clock_year_month_day'
set_microsecond(x, value, ...)

## S3 method for class 'clock_year_month_day'
set_nanosecond(x, value, ...)

Arguments

x
[clock_year_month_day]

A year-month-day vector.

value
[integer / "last"]

The value to set the component to.

For set_day(), this can also be "last" to set the day to the last day of the month.

...
These dots are for future extensions and must be empty.

Value

x with the component set.

Examples

x <- year_month_day(2019, 1:3)

# Set the day
set_day(x, 12:14)

# Set to the "last" day of the month
set_day(x, "last")

# Set to an invalid day of the month
invalid <- set_day(x, 31)
invalid

# Then resolve the invalid day by choosing the next valid day
invalid_resolve(invalid, invalid = "next")

# Cannot set a component two levels more precise than where you currently are
try(set_hour(x, 5))

[Package clock version 0.3.0 Index]