year-month-weekday-group {clock}	R Documentation

Grouping: year-month-weekday

Description

This is a year-month-weekday method for the calendar_group() generic.

Grouping for a year-month-weekday object can be done at any precision except for "day", as long as x is at least as precise as precision.

Usage

## S3 method for class 'clock_year_month_weekday'
calendar_group(x, precision, ..., n = 1L)

Arguments

x
[clock_year_month_weekday]

A year-month-weekday vector.

precision
[character(1)]

One of:

"year"

"month"

"hour"

"minute"

"second"

"millisecond"

"microsecond"

"nanosecond"

...
These dots are for future extensions and must be empty.

n
[positive integer(1)]

A single positive integer specifying a multiple of precision to use.

Details

Grouping by "day" is undefined for a year-month-weekday because there are two day fields, the weekday and the index, and there is no clear way to define how to group by that.

Value

x grouped at the specified precision.

Examples

x <- year_month_weekday(2019, 1:12, clock_weekdays$sunday, 1, 00, 05, 05)
x

# Group by 3 months - drops more precise components!
calendar_group(x, "month", n = 3)

[Package clock version 0.3.0 Index]