Parse into a year-month-day calendar

year_month_day_parse() parses a character vector into a year-month-day calendar.

year_month_day_parse(
  x,
  ...,
  format = NULL,
  precision = "day",
  locale = clock_locale()
)

Arguments

x
[character]

A character vector to parse.

...
These dots are for future extensions and must be empty.

format
[character / NULL]

A format string.

Multiple format strings can be supplied, and they will be tried in the order they are provided.

If NULL, a default format string is chosen based on the precision. The default format string is chosen to parse the result of calling format() on a year-month-day object. For example, with precision = "month", format would be set to "%Y-%m, and with precision = "millisecond" it would be set to "%Y-%m-%d %H:%M:%S".

precision
[character(1)]

The precision of the resulting year-month-day.

locale
[clock_locale]

A locale object created by clock_locale().

Value

A year-month-day calendar vector. If a parsing fails, NA is returned.

Examples

x <- "2019-01-01"

# Default parses at day precision
year_month_day_parse(x)

#> <year_month_day<day>[invalid=0][1]>
#> [1] "2019-01-01"

# Can parse at less precise precisions too
year_month_day_parse(x, precision = "month")

#> <year_month_day<month>[invalid=0][1]>
#> [1] "2019-01"

year_month_day_parse(x, precision = "year")

#> <year_month_day<year>[invalid=0][1]>
#> [1] "2019"

# Even invalid dates can be round-tripped through format<->parse calls
invalid <- year_month_day(2019, 2, 30)
year_month_day_parse(format(invalid))

#> <year_month_day<day>[invalid=1][1]>
#> [1] "2019-02-30"

# Can parse with time of day
x <- year_month_day(
  2019, 01, 30, 02, 30, 00, 5000,
  subsecond_precision = "nanosecond"
)

year_month_day_parse(format(x), precision = "nanosecond")

#> <year_month_day<nanosecond>[invalid=0][1]>
#> [1] "2019-01-30 02:30:00.000005000"

# Can parse using multiple format strings, which will be tried
# in the order they are provided
x <- c("2019-01-01", "2020-01-01", "2021/2/3")
formats <- c("%Y-%m-%d", "%Y/%m/%d")
year_month_day_parse(x, format = formats)

#> <year_month_day<day>[invalid=0][3]>
#> [1] "2019-01-01" "2020-01-01" "2021-02-03"

# Can parse using other format tokens as well
year_month_day_parse(
  "January, 2019",
  format = "%B, %Y",
  precision = "month"
)

#> <year_month_day<month>[invalid=0][1]>
#> [1] "2019-01"

# Parsing a French year-month-day
year_month_day_parse(
  "octobre 1, 2000",
  format = "%B %d, %Y",
  locale = clock_locale("fr")
)

#> <year_month_day<day>[invalid=0][1]>
#> [1] "2000-10-01"