Retrieve low-level zoned time information

zoned_info() retrieves a set of low-level information generally not required for most date-time manipulations. It returns a data frame with the following columns:

begin, end: Zoned times specifying the range of the current daylight saving time rule. The range is a half-open interval of [begin, end).

offset: A second precision duration specifying the offset from UTC.

dst: A logical vector specifying if daylight saving time is currently active.

abbreviation: The time zone abbreviation in use throughout this begin to end range.

zoned_info(x)

Arguments

x
[zoned_time]

A zoned time.

Value

A data frame of low level information to retrieve information for.

Details

If there have never been any daylight saving time transitions, the minimum supported year value is returned for begin (typically, a year value of -32767).

If daylight saving time is no longer used in a time zone, the maximum supported year value is returned for end (typically, a year value of 32767).

Examples

x <- year_month_day(2021, 03, 14, c(01, 03), c(59, 00), c(59, 00))
x <- as_naive(x)
x <- as_zoned(x, "America/New_York")

# x[1] is in EST, x[2] is in EDT
x

#> <zoned_time<second><America/New_York>[2]>
#> [1] "2021-03-14 01:59:59-05:00" "2021-03-14 03:00:00-04:00"

info <- zoned_info(x)
info

#>                                         begin
#> 1 2020-11-01 01:00:00-05:00[America/New_York]
#> 2 2021-03-14 03:00:00-04:00[America/New_York]
#>                                           end offset   dst abbreviation
#> 1 2021-03-14 03:00:00-04:00[America/New_York] -18000 FALSE          EST
#> 2 2021-11-07 01:00:00-05:00[America/New_York] -14400  TRUE          EDT

# `end` can be used to iterate through daylight saving time transitions
# by repeatedly calling `zoned_info()`
zoned_info(info$end)

#>                                         begin
#> 1 2021-03-14 03:00:00-04:00[America/New_York]
#> 2 2021-11-07 01:00:00-05:00[America/New_York]
#>                                           end offset   dst abbreviation
#> 1 2021-11-07 01:00:00-05:00[America/New_York] -14400  TRUE          EDT
#> 2 2022-03-13 03:00:00-04:00[America/New_York] -18000 FALSE          EST