zoned_time_now {clock}	R Documentation

What is the current zoned-time?

Description

zoned_time_now() returns the current time in the corresponding zone. It is a wrapper around sys_time_now() that attaches the time zone.

Usage

zoned_time_now(zone)

Arguments

zone
[character(1)]

A time zone to get the current time for.

Details

The time is returned with a nanosecond precision, but the actual amount of data returned is OS dependent. Usually, information at at least the microsecond level is returned, with some platforms returning nanosecond information.

Value

A zoned-time of the current time.

Examples

x <- zoned_time_now("America/New_York")

[Package clock version 0.3.0 Index]