calendar_group {clock}	R Documentation

Group calendar components

Description

calendar_group() groups at a multiple of the specified precision. Grouping alters the value of a single component (i.e. the month component if grouping by month). Components that are more precise than the precision being grouped at are dropped altogether (i.e. the day component is dropped if grouping by month).

Each calendar has its own help page describing the grouping process in more detail:

year-month-day

year-month-weekday

iso-year-week-day

year-quarter-day

year-day

Usage

calendar_group(x, precision, ..., n = 1L)

Arguments

x
[calendar]

A calendar vector.

precision
[character(1)]

A precision. Allowed precisions are dependent on the calendar used.

...
These dots are for future extensions and must be empty.

n
[positive integer(1)]

A single positive integer specifying a multiple of precision to use.

Value

x grouped at the specified precision.

Examples

# See the calendar specific help pages for more examples
x <- year_month_day(2019, c(1, 1, 2, 2, 3, 3, 4, 4), 1:8)
x

# Group by two months
calendar_group(x, "month", n = 2)

# Group by two days of the month
calendar_group(x, "day", n = 2)

[Package clock version 0.3.0 Index]