clock_locale {clock}	R Documentation

Create a clock locale

Description

A clock locale contains the information required to format and parse dates. The defaults have been chosen to match US English. A clock locale object can be provided to format() methods or parse functions (like year_month_day_parse()) to override the defaults.

Usage

clock_locale(labels = "en", decimal_mark = ".")

Arguments

labels
[clock_labels / character(1)]

Character representations of localized weekday names, month names, and AM/PM names. Either the language code as string (passed on to clock_labels_lookup()), or an object created by clock_labels().

decimal_mark
[character(1)]

Symbol used for the decimal place when formatting sub-second date-times. Either "," or ".".

Value

A "clock_locale" object.

Examples

clock_locale()
clock_locale(labels = "fr")

[Package clock version 0.3.0 Index]